model working with useContext
