+++ src/components/model-table/interfaces.ts
