+++ src/pages/model.page.tsx
@@ -1,32 +1,9 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState, FC } from 'react';
 import { useParams } from 'react-router';
 import { getModelCount, getModelItems, getModelMetadata } from 'services/strapi.service';
 import { ModelTable } from '../components/model-table/model-table';
 import { ModelTablePagination } from '../components/model-table/model-table-pagination';
 import { ModelContext } from '../contexts/ModelContext';
-
-interface Pagination<T = []> {
-    index: number;
-    size: number;
-    count: number;
-    total: number;
-    data: T;
-}
-
-const initialPaginate: Pagination = {
-    index: 0,
-    size: 10,
-    count: 0,
-    total: 0,
-    data: [],
-};
-
-export const usePagination = () => {
-    const [pagination, setPagination] = useState<Pagination>(initialPaginate);
-    // const total = totalPages(modelCount.data.count, pagination.size);
-
-    return { pagination, setPagination };
-};
 
 export const useModel = (contentType: string) => {
     const modelContext = useContext(ModelContext);
@@ -58,7 +35,7 @@ export const useModel = (contentType: string) => {
     };
 };
 
-export const ModelPage = () => {
+export const ModelPage: FC<{}> = () => {
     const { contentType } = useParams<{ contentType: string }>();
     const { metadata, count } = useModel(contentType);
 
@@ -67,12 +44,16 @@ export const ModelPage = () => {
     }
 
     return (
-        <div>
+        <div className="p-4 pl-0">
             <h1 className="mb-1 leading-none text-color-900 text-4xl font-light capitalize">{contentType}</h1>
             <p className="mt-0 mb-4 text-gray-600">{count} entries found</p>
-            <hr className="my-8 border-b-2 border-gray-200" />
-            <ModelTable />
-            {/* <ModelTablePagination totalPages={totalPages(pagination)} /> */}
+            {count > 0 && (
+                <>
+                    <hr className="my-8 border-b-2 border-gray-200" />
+                    <ModelTable />
+                    <ModelTablePagination />
+                </>
+            )}
         </div>
     );
 };
